SR-IOV: macvtap assigned vf check using sysfs
This reverts commits:
 9a022e7d7b85b7c21cf26698fe59c818c4577194
 d7604169988726d121cdc9727accfeb6e29f4aed

As the issue is relevant only for old kernels and almost 4 years
have passed since this fix was introduced, it is safe to undo the
workaround to use ip link show command for determining whether or not
a VF is assigned with macvtap(and followup patch that fixed excessive
use of ip link show commans).

reverting this commit will benefit the agent by reducing the amount of
ip link calls.

While it is possible to perform a revert-per-commit, this change should
really go in as one commit for clarity and avoid introducing a
performance degradation in case one is merged without the other.

 Conflicts:
	neutron/plugins/ml2/drivers/mech_sriov/agent/eswitch_manager.py
	neutron/plugins/ml2/drivers/mech_sriov/agent/pci_lib.py
	neutron/tests/unit/plugins/ml2/drivers/mech_sriov/agent/test_eswitch_manager.py
	neutron/tests/unit/plugins/ml2/drivers/mech_sriov/agent/test_pci_lib.py

 Conflicts details:
   In eswitch_manager: merge tool was unable to properly revert
                       is_assigned_vf() due to changes in later patches.
   In pci_lib: conflicts with later changes to is_macvtap_assigned()
               and link_show() that are no longer required and revert
               of IPCommandDeviceError definition.
   In unit tests: conflicts due to newly added tests and reverting the
                  IPCommandDeviceError exception definition.
                  Sorting out mocks as link_show() method was removed

Related-bug: #1523083

Change-Id: I04ea8eb63de07a6e1e51c2790c5920b086b8542c
(cherry picked from commit 76de8a715d0f48b24fccc06459e9413b96a0449c)